add browser test with zuul
