+++ package.json
@@ -22,14 +22,16 @@
   "homepage": "http://github.com/kevinbeaty/any-promise",
   "dependencies": {},
   "devDependencies": {
-    "promise": "^7.0.0",
-    "es6-promise": "^3.0.0",
-    "rsvp": "^3.0.0",
     "bluebird": "^3.0.0",
-    "when": "^3.0.0",
-    "q": "^1.0.0",
+    "es6-promise": "^3.0.0",
+    "mocha": "^2.0.0",
     "native-promise-only": "^0.8.0",
+    "phantomjs-prebuilt": "^2.1.7",
+    "promise": "^7.0.0",
     "promises-aplus-tests": "^2.0.0",
-    "mocha": "^2.0.0"
+    "q": "^1.0.0",
+    "rsvp": "^3.0.0",
+    "when": "^3.0.0",
+    "zuul": "^3.10.1"
   }
 }
