+++ test/browser.js
@@ -1,0 +1,5 @@
+'use strict';
+
+require('../register')('bluebird', {Promise: require('bluebird')});
+
+require('./tests')
