+++ test/tests.js
@@ -1,30 +1,7 @@
 "use strict";
-var expectedImpl
-if(process.env.ANY_PROMISE){
-  // should load registration regardless
-  expectedImpl = process.env.ANY_PROMISE
-  require('../register')(expectedImpl)
-} else {
-  var version = (/v(\d+)\.(\d+)\.(\d+)/).exec(process.version)
-  if(version && +version[1] == 0 && +version[2] < 12){
-    // Node < 0.12 should load first successful require in
-    // priority list if not registered
-    expectedImpl = 'es6-promise'
-  } else {
-    // Node >= 0.12 should load global.Promise if not registered
-    expectedImpl = 'global.Promise'
-  }
-}
 
 var Prom = require('../');
 var tests = require('promises-aplus-tests');
-var impl = require('../implementation')
-
-if(impl !== expectedImpl){
-  throw new Error('Expecting '+expectedImpl+' got '+impl)
-}
-
-console.log('Starting tests with implementation '+impl);
 
 function deferred(){
   var resolve, reject;
